Rename opentracing.ext.scope_managers to opentracing.scope_managers
