+++ opentracing/mocktracer/tracer.py
@@ -24,7 +24,7 @@ import time
 import opentracing
 from opentracing import Format, Tracer
 from opentracing import UnsupportedFormatException
-from opentracing.ext.scope_managers import ThreadLocalScopeManager
+from opentracing.scope_managers import ThreadLocalScopeManager
 
 from .context import SpanContext
 from .span import MockSpan
