+++ opentracing/scope_managers/__init__.py
