+++ opentracing/scope_managers/constants.py
