+++ testbed/test_active_span_replacement/test_tornado.py
@@ -3,7 +3,7 @@ from __future__ import print_function
 from tornado import gen, ioloop
 
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.tornado import TornadoScopeManager, \
+from opentracing.scope_managers.tornado import TornadoScopeManager, \
         tracer_stack_context
 from ..testcase import OpenTracingTestCase
 from ..utils import stop_loop_when
