+++ testbed/test_client_server/test_asyncio.py
@@ -6,7 +6,7 @@ import asyncio
 import opentracing
 from opentracing.ext import tags
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.asyncio import AsyncioScopeManager
+from opentracing.scope_managers.asyncio import AsyncioScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import get_logger, get_one_by_tag, stop_loop_when
