+++ testbed/test_client_server/test_gevent.py
@@ -7,7 +7,7 @@ import gevent.queue
 import opentracing
 from opentracing.ext import tags
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.gevent import GeventScopeManager
+from opentracing.scope_managers.gevent import GeventScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import get_logger, get_one_by_tag
