+++ testbed/test_common_request_handler/test_asyncio.py
@@ -6,7 +6,7 @@ import asyncio
 
 from opentracing.ext import tags
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.asyncio import AsyncioScopeManager
+from opentracing.scope_managers.asyncio import AsyncioScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import get_logger, get_one_by_operation_name, stop_loop_when
 from .request_handler import RequestHandler
