+++ testbed/test_common_request_handler/test_gevent.py
@@ -4,7 +4,7 @@ import gevent
 
 from opentracing.ext import tags
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.gevent import GeventScopeManager
+from opentracing.scope_managers.gevent import GeventScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import get_logger, get_one_by_operation_name
 from .request_handler import RequestHandler
