+++ testbed/test_listener_per_request/test_tornado.py
@@ -6,7 +6,7 @@ from tornado import gen, ioloop
 
 from opentracing.ext import tags
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.tornado import TornadoScopeManager
+from opentracing.scope_managers.tornado import TornadoScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import get_one_by_tag
