+++ testbed/test_multiple_callbacks/test_gevent.py
@@ -5,7 +5,7 @@ import random
 import gevent
 
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.gevent import GeventScopeManager
+from opentracing.scope_managers.gevent import GeventScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import get_logger
