+++ testbed/test_nested_callbacks/test_asyncio.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 import asyncio
 
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.asyncio import AsyncioScopeManager
+from opentracing.scope_managers.asyncio import AsyncioScopeManager
 from ..testcase import OpenTracingTestCase
 from ..utils import stop_loop_when
