+++ testbed/test_nested_callbacks/test_gevent.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 import gevent
 
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.gevent import GeventScopeManager
+from opentracing.scope_managers.gevent import GeventScopeManager
 from ..testcase import OpenTracingTestCase
