+++ testbed/test_subtask_span_propagation/test_gevent.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, print_function
 import gevent
 
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.gevent import GeventScopeManager
+from opentracing.scope_managers.gevent import GeventScopeManager
 from ..testcase import OpenTracingTestCase
