+++ testbed/test_subtask_span_propagation/test_tornado.py
@@ -5,7 +5,7 @@ import functools
 from tornado import gen, ioloop
 
 from opentracing.mocktracer import MockTracer
-from opentracing.ext.scope_managers.tornado import TornadoScopeManager, \
+from opentracing.scope_managers.tornado import TornadoScopeManager, \
         tracer_stack_context
 from ..testcase import OpenTracingTestCase
